requirements.txt
Flask==2.3.3
gunicorn==21.2.0
torch==2.0.1
numpy==1.24.3
sentence-transformers==2.2.2
transformers==4.33.2
scikit-learn==1.3.0
pandas==2.0.3
faiss-cpu==1.7.4
requests==2.31.0
Werkzeug==2.3.7
gdown
python-dotenv
qdrant-client
tqdm